sdpa fwd (frost): dense padded-Q trim on every SM100 quantized kernel; the fp8 / mxfp8 rows claim it
FlashInfer's dense batched prefill hands per-batch Q lengths with every padded
graph (cuDNN pairs seq_len_q with seq_len_kv under the padding mask). The SM100
fp8 / mxfp8 kernels trimmed padded Q rows only on the d256 / d512 flavors, so the
d128 and (192, 128) flavors either ran a host read (seq_len_q.min().item(),
which breaks CUDA-graph capture) to confirm every length was full, or were
declined. The four remaining kernels now carry the same trim as prefill_d256_fp8:
bounds_for_tile_qtrim collapses the KV loop for CGA tiles past a batch's Q
length, the epilogue writes O := 0 / LSE := -inf on rows past it (folded into
row_dead so the store's select sanitizes NaN padding), the (B,) int32 lengths
ride a dedicated kernel parameter that folds out when the specialization is off.
prefill_d512_fp8 already had the trim but parked the slot after the THD lens
tensors; it sits directly after amax_o everywhere now, where the adapter passes
it. The shared helpers accept a raw arith batch index (the d192 fp8 correction
loop hands one). Rows: sdpa_fwd_prefill_sm100_fp8 dense_seq_q_trim=not rubin_row,
sdpa_fwd_prefill_sm100_mxfp8 dense_seq_q_trim=True; the adapter's
_quantized_q_lens_abi is every SM100 quantized flavor. Tests: the fp8 / mxfp8
dense-Q-trim tests run on d128, (192, 128), d256 (and d512 for mxfp8).

+++ python/cudnn/sdpa/fwd/kernels/_common_blackwell.py
@@ -695,7 +695,7 @@ def make_sdpa_helpers(
             cga_base_super = q_super_idx - cta_in_pair
             q_row_coord = cga_base_super * cutlass.Int32(tokens_per_super)
             arr = cutlass.make_array_view(seq_q_lens_tensor)
-            q_len_b = cutlass.Int32(arr[batch_idx])
+            q_len_b = cutlass.Int32(arr[cutlass.Int32(batch_idx)])  # batch_idx may arrive as a raw arith value from the decode
             tile_dead = q_row_coord >= q_len_b
             dead_lo = cutlass.Int32(arith.select(tile_dead.ir_value(), b.left.ir_value(), b.unmasked_lo.ir_value()))
             dead_hi = cutlass.Int32(arith.select(tile_dead.ir_value(), b.left.ir_value(), b.unmasked_hi.ir_value()))
@@ -731,7 +731,7 @@ def make_sdpa_helpers(
             return cutlass.Int32(cu[q0 + batch_idx + cutlass.Int32(1)]) - cutlass.Int32(cu[q0 + batch_idx])
         if cutlass.const_expr(int(getattr(CFG, "SEQ_Q_LENS_PRESENT", 0)) == 1 and int(CFG.BOTTOM_RIGHT) == 1):
             arr = cutlass.make_array_view(seq_q_lens_tensor)
-            return cute.math.max(cutlass.Int32(0), cute.math.min(cutlass.Int32(arr[batch_idx]), scalar_seqlen_q))
+            return cute.math.max(cutlass.Int32(0), cute.math.min(cutlass.Int32(arr[cutlass.Int32(batch_idx)]), scalar_seqlen_q))
         return scalar_seqlen_q
 
     _thd_on = int(getattr(CFG, "THD_VARLEN", 0))
